Etnmine.pro
Could you please add the following pool into your pool list? At the moment the pool is running one node in London(UK) and another one in Nuremberg (Germany). We are a small team of developers who run this as a side project, so expect a high-availability, low latency and a lot more ETN nodes.